refactor: Nickname Input Page
