+++ frontend-hybrid/yakal/lib/screens/Login/NicknameInput/style.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+import '../../../utilities/style/color_styles.dart';
+
+class NicknameInputStyle {
+  static const title = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w500,
+    height: 1.4,
+  );
+
+  static const boldTitle = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w500,
+    height: 1.4,
+  );
+
+  static const label = TextStyle(
+    color: ColorStyles.gray4,
+    fontSize: 15,
+    fontWeight: FontWeight.w600,
+    height: 1.6,
+  );
+
+  static const inputHint = TextStyle(
+    color: ColorStyles.gray3,
+    fontSize: 15,
+    fontWeight: FontWeight.w500,
+    height: 1.6,
+  );
+
+  static const input = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 20,
+    fontWeight: FontWeight.w600,
+    height: 1.2,
+  );
+
+  static const error = TextStyle(
+    color: ColorStyles.red,
+    fontSize: 16,
+    fontWeight: FontWeight.w700,
+    height: 1.6,
+  );
+}
